backend/requirements.txt
# Web Framework
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-login==0.6.3
flask-cors==4.0.1
firebase-admin==6.5.0

# Database
sqlalchemy==2.0.30

# Resume Parsing
PyPDF2==3.0.1
python-docx==1.1.2

# NLP & Skill Matching
spacy==3.7.5
scikit-learn==1.4.2
numpy==1.26.4
flashtext==2.7

# Job API integration
aiohttp==3.9.5
requests==2.32.3
beautifulsoup4==4.12.3

# AI Resume Generation
google-generativeai==0.7.2

# Utilities
python-dotenv==1.0.1
werkzeug==3.0.3
markdown==3.6
